support for installing on apple silicon via homebrew (#1264)

+++ update_homebrew_formula.sh
@@ -3,6 +3,7 @@ set -e
 
 VERSION=$1
 SHA=$2
+SHA_ARM=$3
 
 if [ -z "$VERSION" ]; then
   echo "missing version"
@@ -14,31 +15,52 @@ if [ -z "$SHA" ]; then
   exit 1
 fi
 
+if [ -z "$SHA_ARM" ]; then
+  echo "missing sha for ARM"
+  exit 1
+fi
+
 rm -rf homebrew-cli
 git clone --depth 1 https://github.com/okteto/homebrew-cli.git
 pushd homebrew-cli
 
 cat << EOF > Formula/okteto.rb
 class Okteto < Formula
-    desc "CLI for cloud native development"
-    homepage "https://okteto.com"
-    url "https://github.com/okteto/okteto/releases/download/$VERSION/okteto-Darwin-x86_64"
+  desc "Develop and test your code directly in Kubernetes"
+  homepage "https://github.com/okteto/okteto"
+  version "$VERSION"
+  license "Apache-2.0"
+  
+  if Hardware::CPU.arm?
+    sha256 "$SHA_ARM"
+    url "https://github.com/okteto/okteto/releases/download/$VERSION/okteto-Darwin-arm64"
+  else
     sha256 "$SHA"
-    version "$VERSION"
-    
-    head do
-        url "https://downloads.okteto.com/cli/master/okteto-Darwin-x86_64"
-    end
-    
-    def install
-        bin.install "okteto-Darwin-x86_64"
-        mv bin/"okteto-Darwin-x86_64", bin/"okteto"
-    end
+    url "https://github.com/okteto/okteto/releases/download/$VERSION/okteto-Darwin-x86_64"
+  end
 
-    # Homebrew requires tests.
-    test do
-        assert_match "okteto version $VERSION", shell_output("#{bin}/okteto version 2>&1", 0)
+  head do
+    if Hardware::CPU.arm?
+      url "https://downloads.okteto.com/cli/master/okteto-Darwin-arm64"
+    else
+      url "https://downloads.okteto.com/cli/master/okteto-Darwin-x86_64"
     end
+  end
+  
+  def install
+    if Hardware::CPU.arm?
+      bin.install "okteto-Darwin-arm64"
+      mv bin/"okteto-Darwin-arm64", bin/"okteto"
+    else
+      bin.install "okteto-Darwin-x86_64"
+      mv bin/"okteto-Darwin-x86_64", bin/"okteto"
+    end
+  end
+
+  # Homebrew requires tests.
+  test do
+      assert_match "okteto version $VERSION", shell_output("#{bin}/okteto version 2>&1", 0)
+  end
 end
 EOF
